Only show links to google/ChatGPT if the site is localhost
